♻️🎪 Extract route loader and action
