+++ apps/show/src/routes/_exhibitor.exposants.$token._config.stand.modifier-stand/form.ts
@@ -3,7 +3,8 @@ import { useForm as useFormBase } from "@conform-to/react";
 import { getZodConstraint, parseWithZod } from "@conform-to/zod";
 import { useActionData, useLoaderData } from "@remix-run/react";
 import { ActionSchema } from "./action";
-import type { action, loader } from "./route";
+import type { action } from "./action.server";
+import type { loader } from "./loader.server";
 
 export function useFormRoot() {
   const { exhibitor } = useLoaderData<typeof loader>();
