+++ apps/show/src/routes/_exhibitor.exposants.$token._config.stand.modifier-stand/route.tsx
@@ -1,50 +1,14 @@
 import { getErrorTitle } from "#core/data-display/error-page";
 import { FormLayout } from "#core/layout/form-layout";
 import { createSocialMeta } from "#core/meta";
-import { Routes } from "#core/navigation";
 import { getPageTitle } from "#core/page-title";
-import { badRequest } from "#core/response.server";
-import { services } from "#core/services.server.js";
-import { RouteParamsSchema } from "#exhibitors/route-params";
-import { safeParseRouteParam } from "@animeaux/zod-utils";
-import { parseWithZod } from "@conform-to/zod";
-import { ShowExhibitorStatus } from "@prisma/client";
-import type { ActionFunctionArgs, LoaderFunctionArgs } from "@remix-run/node";
-import { json, redirect } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/react";
-import { ActionSchema } from "./action";
+import type { loader } from "./loader.server";
 import { SectionForm } from "./section-form";
 import { SectionHelper } from "./section-helper";
 
-export async function loader({ params }: LoaderFunctionArgs) {
-  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
-
-  const exhibitor = await services.exhibitor.getByToken(routeParams.token, {
-    select: {
-      name: true,
-      activityFields: true,
-      chairCount: true,
-      dividerCount: true,
-      dividerType: true,
-      hasElectricalConnection: true,
-      hasTablecloths: true,
-      installationDay: true,
-      peopleCount: true,
-      placementComment: true,
-      size: true,
-      standConfigurationStatus: true,
-      tableCount: true,
-      updatedAt: true,
-      zone: true,
-    },
-  });
-
-  if (exhibitor.standConfigurationStatus === ShowExhibitorStatus.VALIDATED) {
-    throw redirect(Routes.exhibitors.token(routeParams.token).stand.toString());
-  }
-
-  return { exhibitor };
-}
+export { action } from "./action.server";
+export { loader } from "./loader.server";
 
 export const meta: MetaFunction<typeof loader> = ({ data }) => {
   return createSocialMeta({
@@ -55,44 +19,6 @@ export const meta: MetaFunction<typeof loader> = ({ data }) => {
     ),
   });
 };
-
-export async function action({ request, params }: ActionFunctionArgs) {
-  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
-
-  const exhibitor = await services.exhibitor.getByToken(routeParams.token, {
-    select: { standConfigurationStatus: true },
-  });
-
-  if (exhibitor.standConfigurationStatus === ShowExhibitorStatus.VALIDATED) {
-    throw badRequest();
-  }
-
-  const formData = await request.formData();
-
-  const submission = parseWithZod(formData, { schema: ActionSchema });
-
-  if (submission.status !== "success") {
-    return json(submission.reply(), { status: 400 });
-  }
-
-  await services.exhibitor.updateStand(routeParams.token, {
-    chairCount: submission.value.chairCount,
-    dividerCount: submission.value.dividerCount,
-    dividerType: submission.value.dividerType,
-    hasElectricalConnection: submission.value.hasElectricalConnection,
-    hasTablecloths: submission.value.hasTablecloths,
-    installationDay: submission.value.installationDay,
-    peopleCount: submission.value.peopleCount,
-    placementComment: submission.value.placementComment || null,
-    size: submission.value.size,
-    tableCount: submission.value.tableCount,
-    zone: submission.value.zone,
-  });
-
-  services.exhibitorEmail.standConfiguration.submitted(routeParams.token);
-
-  throw redirect(Routes.exhibitors.token(routeParams.token).stand.toString());
-}
 
 export default function Route() {
   return (
